Error fix

+++ src/ft_tst_util2.c
@@ -6,15 +6,15 @@
 /*   By: hojsong <hojsong@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 23:00:02 by hojsong           #+#    #+#             */
-/*   Updated: 2024/04/13 04:19:48 by hojsong          ###   ########.fr       */
+/*   Updated: 2024/04/17 16:41:00 by hojsong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/malloc.h"
 
 t_sta	*g_all;
-int		g_blakcs_size;
-int		g_blakc_one;
+size_t	g_blakcs_size;
+size_t	g_blakc_one;
 
 t_st	*t_stinit(t_st *src, size_t size)
 {
