+++ src/ft_tst_util4.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_malloc.c                                        :+:      :+:    :+:   */
+/*   ft_tst_util4.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hojsong <hojsong@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 22:06:12 by hojsong           #+#    #+#             */
-/*   Updated: 2024/04/13 05:21:55 by hojsong          ###   ########.fr       */
+/*   Updated: 2024/04/17 16:45:35 by hojsong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/malloc.h"
 
 t_sta	*g_all;
-int		g_blakcs_size;
-int		g_blakc_one;
+size_t	g_blakcs_size;
+size_t	g_blakc_one;
 
 void    init(void)
 {
