+++ src/show_util.c
@@ -6,15 +6,15 @@
 /*   By: hojsong <hojsong@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/13 00:40:53 by hojsong           #+#    #+#             */
-/*   Updated: 2024/04/13 04:03:31 by hojsong          ###   ########.fr       */
+/*   Updated: 2024/04/17 16:41:07 by hojsong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/malloc.h"
 
 t_sta	*g_all;
-int		g_blakcs_size;
-int		g_blakc_one;
+size_t	g_blakcs_size;
+size_t	g_blakc_one;
 
 int	m_ft_strcmp(char *str1, char *str2)
 {
